+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod control;
 pub mod engine;
 pub mod metadata;
+mod mp4;
 pub mod planner;